WIP, working locally, need network code refactor.

+++ point_ao_split_rendering/PointHashGenerator.cpp
@@ -17,7 +17,7 @@ void PointHashGenerator::generate(
   for (uint32_t instanceId = 0; instanceId < scene->getGeometryInstanceCount(); instanceId++) {
     // NOTE: sometimes the hash table size can be a bit too small, and ideally we want to configure
     // it and/or figure out dynamically how much we need
-    uint32_t hashTableSize = std::exp2(
+    uint32_t hashTableSize = (uint32_t)std::exp2(
         std::ceil(std::log2((numFinalSamplesPerInstance[instanceId] / FIXED_HASH_BUCKET_SIZE))) +
         HASH_LOG2_SIZE_FACTOR);
